Refactor common functions to utility file
